Migrate search to new input system
Finally we can start to use Misa's overhauled input system. Earlier I
tried to migrate the script editor to it, but I got kinda entangled in
that because the input system was new to me and I didn't decide to
start small - with a portion of Ved that didn't have as many entry
points and couldn't do a lot of different things with the input while
it was still going to keep being displayed or while switching to a
different input text.

Either way, I started small this time so we can see how it goes and
fix possible bugs and design problems before doing the tough complex
inputs.

[pre10]