Split main2.lua into different files
main2 just had all the LÖVE callbacks in one place, but it's probably
better if each callback goes into its own file. This should help make
things easier to find, and main2 is a pretty weird name anyway - it was
just that main.lua (what LÖVE requires you to start with) already has
its own preparation code to determine if Ved is able to run at all on
your version of LÖVE and shouldn't show an alternative screen for
example.

There was also a list of states and debug shortcuts at the top of
main2. The shortcuts list can go, but I moved the states list to a file
states.txt, to make sure there's always an authoritative list in any
specific version or branch.

[pre01]