Restructure UI files
UI code (in uis/) is now no longer one file per UI, but one folder per
UI. All the UI's callbacks and the elements table are now in dedicated
files, which are all loaded in automatically by load_ui(ui_name) if
they exist.

Re-specifying the name of the UI in the UI's code (so it could be
reloaded later if needed) is now also no longer necessary. But in this
style I was briefly thinking all it would need is a name.lua which
returns a string.

This is in line with 4b9899d276e7c5b6ec9bc602d3e2a281845c6fb9, where
main2.lua was split into different files to help make things more
readable.

[pre03]