Remove draw*() functions in draw*.lua files
Files like drawmaineditor.lua or drawmap.lua were dedicated to the long
drawing functions for their applicable states. Some of these were added
very early on, because having all the drawing code for the main editor
in the middle of an `if state == 1 then`, inside love.draw(), inside
main2.lua was unwieldy and already beyond ugly at the time.

Now these files are redundant as well, since uis/NAME/draw.lua is now a
thing. With the UI file split I don't have to be hesitant to move the
drawing code for states out of their dedicated files anymore, because it
won't drown out the other UI callbacks.

[pre03]