Switch update check to NSURLSession instead of LuaSocket on macOS
Also made a consistent way to make an HTTPS request on all platforms,
require https_main.lua and call https_request(url).

The plain HTTP LuaSocket code is now completely removed. This also
means platforms that Ved doesn't officially support, like Android and
iOS, can't check the version anymore. That shouldn't be too big a deal
though, especially since LÖVE 12 will probably come with built-in HTTPS
support, so that's gonna be a nice alternative for those platforms, and
also for Linux, to avoid the update check depending on wget on LÖVE 12.

Then about macOS, I'm happy that I finally understand Objective-C's
syntax, and that I can now add Objective-C to the list of programming
languages I've worked with.

This is tested on OS X 10.9 (Mavericks) (can't get tolp.nl's
certificate to be accepted on that old system though, but at least I
tested the code itself to work with a different site and it doesn't
crash in either case) and on 10.13 (High Sierra). May also get the
chance to test it on more recent versions as well, at least 10.14
(Mojave).

Possible todo: ensure it doesn't break in 10.8 and below. NSURLSession
was added in 10.9, and it's probably not that hard to check the OS
version and bail out early.

[pre04]