Remove VCE support
VCE is dead. So there's no reason for Ved to support it anymore.

This still keeps all refactors necessary to support multiple versions of
VVVVVV within Ved, which is mostly just the `limit` stuff in the
metadata. `metadata.target` will still be kept, and it will always be
"V", and all `metadata.target == "V"` checks are still kept.