Get rid of most Whitesmiths-like table indentation
So instead of:

thing =
	{
	"a",
	"b"
	}

We use something like this instead:

thing = {
	"a",
	"b"
}

I basically self-invented (a variant of) Whitesmiths back in the day
and felt it was more logical, but the world disagreed so much I got
used to other styles years ago. Maybe some of it was also added later
to be consistent with the other Whitesmiths code. Anyway, it should be
changed on sight now.

[pre07]