Use `DFP.` constants for (numeric) dialog field property keys
Dialog fields have properties (like their key, position, current value,
but also specific ones for certain types like dropdown menu items,
onchange handlers, file picker error messages...), and these properties
are all stored sequentially in the fields. I now think this was a poor
design choice, especially now that the fields have become more complex
(so to be fair, the numeric keys were already a thing when you couldn't
move the cursor in text input fields and the only extended options that
existed were for dropdowns).

The biggest problem was that the code had to work with these numeric
properties too, picture appending a field's [7] to that field's [5],
or if the [6] was DF.FILES and we select an item from [7], we set the
[10] to 0, and make sure the [11] is updated... You get the idea.

The main thing I want to do is to just use proper names for the
properties, such as field.value instead of field[5], but since it seems
that would require some pretty nontrivial changes, I've just taken away
the biggest readability problem by replacing the numbers by constant
values in a table DFP. wherever I could find them. So now it looks like
field[DFP.VALUE] or field[DFP.FILES_MENUITEMS], which is not perfect,
but it's much clearer already.

[pre07]