Add VVVVVV-CE deprecation warning
It now warns you if you convert a level from VVVVVV to VVVVVV-CE, or if
you open a VVVVVV-CE level.

Also found and removed a remnant of the new VVVVVV-CE tools that would
cause level comparisons to speak about entities of type $1 instead of
something like type 8, now that the strings for the tools have been
removed.

[pre09]