Fix clicking on Unicode characters in the script editor
I think the original rationale for not doing it properly was that
Unicode inside scripts was pointless anyway since VVVVVV didn't support
it at all, so typing Unicode was also blocked in most of Ved.

Never mind that pasting it wasn't... But nobody would ever do such a
thing in a script because it'd be a waste of time.

VVVVVV 2.3 does support displaying Unicode, but even if it didn't, Ved
shouldn't just crash if you click on a Unicode character in a script.

[pre00]