Ignore data changes from non-text entities in level diffs
It is possible for entities to be marked as changed in level diffs even
when they haven't been changed at all. This is because their data
attribute has been changed, which happens due to the bug with the
in-game level editor where, upon saving, the data attributes of non-text
entities are filled with random strings from memory. (It is assumed that
this bug is fixed in 2.3.)

That bug results in the entity diffs of the 2.1 and 2.0 versions of A
New Dimension being extremely noisy.

To fix this, just ignore data attribute changes if it comes from a
non-text entity. Now the entity diffs of A New Dimension 2.1 and 2.0 are
no longer filled with noise.