Clarify playtesting error messages, add error messages for Linux/macOS
The Linux and macOS code had no error checking for playtesting and
always reported success, now it at least checks whether piping data
went successfully (since we don't really seem to be able to check
whether the process was created successfully, but we can check whether
data was successfully sent to a process).

When an error does occur, there's more information, instead of just
the OS message thrown in a box.

[pre02]