Reduce likelihood of crashes due to invalid tiles/sprites size
If the tileset has dimensions that are not multiples of 8, VVVVVV 2.3
gives an error (and closes) while Ved would crash. Now Ved floors the
calculated number of columns and rows in the tilesets and spriteset (so
partial tiles will not be included at all), which is technically
inconsistent with VVVVVV but it's better than a crash. This shouldn't
be happening anyway; most people would probably expect tilesets to have
to be left as exactly their original size and expect problems if they
aren't.

[pre02]