Add alignment support to list containers
Previously, elements in ListContainers and HorizontalListContainers
would always be horizontally/vertically centered. Now there's an
argument that can change this to left/right/top/bottom. This argument
is conveniently added right in the middle of the others which is
technically a breaking API change, but it's worth it because it's a
much more logical place than at the end, and it's not like there's
masses of plugins using list container elements. And even then, it's
documented in the "Changelog of breaking codebase changes".

[pre03]