Add "VVVVVV setup" options screen
This new screen replaces the current VVVVVV folder dialog, and puts it
in a dedicated page of the settings. This screen will later also have
controls to set/change the VVVVVV executable used for playtesting, to
hopefully take away some of the unclarity how to change it once it's
set (more clarity in the main editor is also planned). It could have
more settings for playtesting as well if needed (I've always been
having some kind of argument customization in mind but maybe that's not
necessary)

Also, finally a state which fully uses the new GUI elements system
instead of just the right bar! No specific drawing functions (or
love.graphics. things) are used at all. I can tell the elements system
will need to go through a lot of growing pains as more and more stuff
uses it and requires small things that were unaccounted for, but we
need to start somewhere.

[pre03]