Replace all instances of `rvnum` by `script_i` (or `article_i` in help)
I don't even know what rvnum is supposed to be an abbreviation of
anymore, or if it's even an English abbreviation. I think it means
"ReVerse NUMber of script" since the loop went from last to first, but
either way, it's always used to mean the index of a script in the
scriptnames table, or in the help system it means the index of an
article (where the order is not even ReVersed), so I decided to replace
it by script_i and article_i as appropriate.

[pre05]