Add right-clicking the playtest button re-asking for VVVVVV executable
It has always been possible to hold Shift while pressing Enter or
pressing the playtest button to change the playtesting path, but that
was not obvious enough.

I was going to add a one-time dialog or some on-screen text about it,
like "In case you need to change the path to VVVVVV, hold Shift while
starting or go to Ved options.", but I just can't be satisfied with it,
that just feels like failed design.

So instead, I just made the playtest button right-clickable, that seems
to me like something people would try, is in line with the custom brush
reset, and is hopefully enough in combination with the new VVVVVV setup
options.

[pre07]