Change help system to not require "Return" article
The return button in the help, level notes, latest info pages, level
differences, etc, has always been an article called "Return" with a
formatting code (that nowadays was hardcoded to only work in the first
article), that changed the state to the previous state. Now the Return
button doesn't lead to an article but actually has an effect itself,
all return articles are gone, and the `)` formatting code is removed as
well.

[pre08]