Clean up nearly all remaining VCE parts
Now there might be some random stuff left like a check whether the
selected tool is 19 via a table.contains or something (I wouldn't have
found that if there wasn't also a `selectedtool == 19` around there)...

But as far as I can tell, this should cover most nooks and crannies.
Now the only item left on my grown todo list (which I kinda want to
share for interest's sake, but it's fully in Dutch) is altstates.

[pre04]