Fix crash in crash screen if crash happens in drawing code with canvas
I've seen it happen recently during development that the crash screen
crashed with a "present cannot be called while a Canvas is active"
error. Turns out it's part of the "let's try love.draw() one more time
so we can freeze the picture" code, which has a love.graphics.present()
explaining the issue.

[pre14]