Remove deprecated shortcuts
This involves all the shortcuts that were deprecated to make room for
either the new input system, or for VCE and because new shortcuts made
more sense than the old ones. They all had a notification pop up with
appropriate strings L.OLDSHORTCUT_*:
- CTRL+left/right instead of ALT+left/right in the script editor
- Ctrl+A instead of Ctrl+R (Resources) for Assets
- Ctrl+D instead of Ctrl+F for opening levels folder
- Q instead of G for going to a room by coordinates
- K instead of Shift+; for hiding room contents

It's already been longer than anticipated for needing to remove these
shortcuts, so it's time to get rid of them. I still find it hard to
break my old habits for some of these, but "everyone" should now be
aware of the changes, and it's probably easier to finally get used to
the new ones if the old ones no longer actually work.

[pre15]