Convert script editor to new input system - mostly
This has been exhausting, but it looks like it's nearly ready!
It still has some known bugs/incompletenesses and it's gonna need
a TON of testing (I practically didn't test it yet), but now the
bulk of the work is done. And it's not broken the way it was last time
where I had to basically give up (branch scripteditor_newinputsys)

scriptlines and editingline are no longer a thing, now obviously
inputs.script_lines and, locally requested where needed, editing_line
are used (switching to a different name also helps with finding what's
left to be done). In other places, it simply makes a copy of the script
to work with instead of directly messing around in the input that might
not actually get closed in all cases.

processflaglabels() and processflaglabelsreverse() now work differently
too, they take the raw and human-readable scripts as arguments
respectively and return the converted script. And I'm still going to
rename those functions.

[pre19]