Fix unused flag reuse bug in script_compile
This is not actually a regression from the changes for the new input
system in the script editor, but I discovered it while testing a script
with 101 new flags. After fiddling with it for a while, the script was
accepted all of a sudden, and it turned out two flags had merged into
one.

If a flag number has been assigned a name before, but it's now unused,
the compile function can reuse it. However, if a flag name that was
assigned to the first unused flag number gets used in a script, and
then a new flag name is used in the same script, it didn't account for
the fact that that flag is now in use again and can end up using the
same number for the two.

Simplest way to reproduce, create this script:
	flag(a,on)

Exit and re-enter the editor, clear it out, exit and re-enter, now make
the script the following:
	flag(a,on)
	flag(b,on)

At this point, both a and b will be number 0.

[pre21]