Add teleportscript as command that references custom_ scripts
teleportscript() doesn't execute scripts directly. However, it still
makes sense to mark the scripts it refers to as referenced, since the
script containing the command needs to be run to reference that script
in the first place, and the script being referred to will then not be
marked as unused if no other scripts reference that script.