Sync internal commands list
This involves knowninternalcommands and the English help page.
audiopause and changerespawncolour are now listed, so they will no
longer show up as red in the script editor.
The help page still needs to be synced across languages and then
translated, but that can be done around tomorrow (and would
happen anyway since it's now in the English language file)

Ticks one item in #133.

[pre24]