Show alternative style of bars for unlimited stats
It's annoying to not have any visual indication on the stats screen for
how many scripts or entities there are in a level since the bar is,
logically, always 0% filled. Now those bars (visually) count towards
the old limit again, indicated with a line and the number. Once that
old limit has been surpassed, the bar stays full, but the old limit
line starts growing to the left to put the value in perspective:

https://tolp.nl/host/unlimited_bars.png

(My previous idea was to define the full width of the bar as twice the
old limit, have a line at half of the bar to indicate the old limit,
and have the bar itself max out at twice this old limit. But the new
idea works with a wider range of values.)

[pre26]