Remove possibility for levelmetadata_get to ever return a second value
This second value was only returned to signal voided metadata, which
belong to out-of-bounds rooms. Which never happens, unless you
specifically ask for it by editing the config file to enable breaking
the limit, and then make a level too big and override the warning
message.

The problem with the second return value is that levelmetadata_get is
used in a LOT of places that expect only a single value to come out
(like as one of the arguments to a function), and if a second value
suddenly does come out, it shifts other arguments out of the way or
suddenly provides a value to an optional argument... Resulting in some
absolutely diabolical bugs that require you to spell out exactly which
values get passed to where in code that looks like it should work.

Anyway, so >20x20 support was broken and has now been fixed. I feel
like it would be better to remove the feature altogether since it needs
complex code (just look at map_correspondreset(), for example) and I
don't think many people have a reason to use it, so it's probably not
worth it. It brought the levelmetadata and roomdata getters and
setters, which are good for code quality, but I think more parts of Ved
should have such interfaces anyway (instead of getting and modifying
data directly) and those functions would also be better without the
special handling.

But for now, easy enough fix is that voided metadata now has an
attribute voided=true.

[pre30]