Add support for font.txt (in graphics folder)
This extends the existing VVVVVV/graphics/font.png option with support
for custom font characters, instead of the regular ASCII. As a bonus,
this integrates really well with language-specific fontpng_ascii(c)
functions, as supporting a character in the font will exempt that
character from being replaced by fontpng_ascii.

This is just the first step to supporting level-specific fonts, next
step I mainly need to worry about keeping the level font state
properly; loading different levels in a row should not reload the font
if none of them have a custom one, but loading a level without a font
after a level with a font should revert back to the built-in font (or
graphics folder font). And about the decision what to do with the
existing setting for optionally enabling font.png.

[pre31]