Add support for per-level custom fonts
Now we can tick another box in #133.

A new config option has been added that controls whether level-specific
fonts should be used, enabled by default. Now we have a disabled-by-
default option for VVVVVV/graphics/font.png and an enabled-by-default
option for level-specific fonts, but that's probably okay; if a level
provides a font then it's probably needed for the level in some way,
or it indicates the level is intended to be viewed with that font, so
it makes sense to view the level with that font by default. A font from
the graphics folder on the other hand would apply to ALL levels, so
it's probably a more cosmetic difference in that case, and it's an
option to view Ved in that font as well.

(Known bug in font.txt handling that I just discovered: the first 32
characters will always be spaced as if they're U+0000 - U+001F, even if
font.txt defines them differently.)

[pre32]