Fix text cursor not moving when entering folder in levels list
Entering a folder would fill in the name of that folder on the input
field, but your cursor would remain in place (now in the middle of the
text). It now moves to the end properly.

I also discovered how "rightmost" works in the input system, so I'm now
using that wherever I used to do setpos with utf8.len.

[pre35]