Improve editor mousepress handling (especially eraser/entity deletion)
The reason for the button to disable the eraser has always been a bit
of a flaw in the UI in my opinion, where having an eraser tool that has
to be able to erase any tile, and having a right click menu on any
entity, clashed with each other. Right clicking an entity would also
erase a tile behind that entity unless the eraser was disabled (or a
non-drawing tool was selected). Erasing tiles and sliding over an
entity's control tile would also interrupt the erasure by bringing up
the right click menu.

Now, right clicking an entity won't erase any tiles, and slide-right-
clicking over an entity's control tile won't interrupt the erasure!

This required some code to be moved around, so I took the chance to do
things in a better way. And incidentally, I think this change has been
pretty good for the organization of the code. Handling of using the
mouse buttons (I'm deliberately not saying "clicking" since that kinda
feels like it implies holding the button down doesn't count) for both
the tools and for right-clicking/moving/copying entities has been taken
out of drawing code, and into their own files and functions that take
place in love.update (pretty appropriate if you have to be able to hold
down the mouse button). In the case of editor tools, this was just a
big block of code that was moved out of the drawing function for the
main editor. In the case of entities, it was more entangled, so the
entity drawing code is now also cleaner, and there's now a function
that only focuses on stuff like right click menus for entities.

The button to disable the eraser can now be removed in my opinion, but
I'll do that in a separate commit tomorrow.

[pre39]