Fix text cursor not being changed to pointer when opening context menu
Apparently this has been a thing since the commit that migrated the
script editor to the new input system. I was surprised by that - I
thought I tested that - and the input system right click menu - plenty
of times and should've noticed the cursor being wrong on the buttons of
the right click menu, but I guess I must've made the change that
caused this bug (jumping out of input.getfocused() if not nodialog)
after I did any such testing.

Now creating a right click menu resets the cursor to default, and
besides, resetting the cursor is now simplified to a function call
(reset_special_cursor().)

[pre40]