Remove support for contents bigger than actual size
Ved's bigger-than-20x20 support has support for tiles in the <contents>
tag that is outside the actual size of the array in VVVVVV (which is
20 * 20 * 40 * 30 = 480000). The main reason I added this support was
because I thought it supported it, but in reality I was actually just
stomping on memory that didn't belong to me, and it just happened to
work because std::vector allocates up to the nearest power of two, and
the memory initialized to 0.

Anyways, 2.3 no longer allows any sort of tiles past the contents limit,
so remove support for this.