Re-add tab autocompletion to the script editor and improve it
The old tab autocompletion only worked if there was only one possible
command you wanted to use, otherwise it did nothing, just in case.
The migration to the new input system disabled tab autocompletion for
a bit, but now it's back, and I decided to make it more usable: it now
always completes to the shortest possible command that matches your
prefix. The idea is that you might press tab earlier for shorter
commands than you might for longer ones (I guess? From the human
perspective of "it should be able to guess what I mean if I gave it
as much as 50% of the command"), short commands may be part of longer
ones anyway (text/textboxactive, speak/speak_active), and if it
completes to the wrong command, there's a minimum number of characters
that you need to delete.

For example, if you type "te", there are two matches, "text" and
"textboxactive", so it will pick "text". In fact, if you tab again,
"textboxactive" is now the shortest (and only) possible completion.

Of course a system where you can always tab multiple times to get
different results would be better, but this is already a good
improvement for now.

[pre42]