Fix a few crashes when editing number values in certain cases
Specifically, when you backspace the numbers in the map export dialog.
This is caused by the switch to consistent use of s:sub(...) instead of
string.sub(s, ...), because if s is a number, s:sub(...) won't work but
string.sub(s, ...) silently will, which is very annoying. It could be
argued I should revert those instances, but that'd only defer the bug
to whenever I decide to refactor those instances again. So I just
tostring'd the text there. The functions in question can probably be
rewritten to be much simpler anyway now that we have the utf8 library
in 0.9.1 as well, and they'll probably become obsolete anyway with the
switch to the new input system.

[pre44]