Fix plugin code changes being double-escaped on changing language/font
Changing the language or font requires UI elements to be reloaded
unfortunately (I should work around that) and that means plugin source
edits have to be re-applied too. The edits sometimes need to be escaped
with percent signs first. But the code that does so pretty much assumed
ved_require(...) can only be called once for a file, and overwrote the
original find/replace strings with escaped versions. So this may cause
problems if a plugin edit to files that need to be reloaded involves
characters that need to be escaped, and you then do change the language
or the font setting.

[pre46]