Fix exported map images being mis-scaled in very specific conditions
This very specific combination of conditions affecting map scale is...
eyebrow-raising, but I swear it makes sense.

The map would be mis-scaled if ALL of these apply at the same time:
- Ved itself is scaled (for example 2x instead of 1x)
- Dialog animations are disabled
- You click the Save button in the "Save map" dialog, instead of
  pressing Enter (the same button's hotkey)

So, LÖVE doesn't seem to have the perfect solution to scaling the
entire window easily (haven't looked into it for a while) and it
requires more manual work to make the existing code work as expected.
Getting the mouse position, the actual size of the window... Anyway, at
the start of love.draw(), we set the scaling that ensures anything
that's drawn to the screen is scaled and positioned properly, and at
the end of love.draw() we reset it again.

Drawing to canvases is also affected by this love.graphics.scale(...).
The map export feature uses a canvas to draw the rooms to, but the code
that draws the rooms on the canvas is outside the scope of love.draw().
(obviously, why wouldn't it be...)

Dialog buttons can be pressed with a possible keyboard hotkey (handled
in the dialog's keypressed callback) or by clicking them. Ved's dialog
system fades out the dialog windows when they disappear, so pressing a
button first fades out the window and only then calls the handler (both
handled in the dialog's update callback).

So, even though clicking a button is checked in the dialog's draw
callback because that's where it determines where each button is in the
first place, the actual handler is called outside of drawing code.

Sike, dialog animations can be turned off, so in that case pressing a
button might as well call the handler right away, which is thus still
in drawing code in the case of a click. This shouldn't affect anything,
of course, but turns out it does, if you happen to need to draw to a
canvas in the dialog handler while scaling is active.

The solution: don't call the dialog handler right away if animations
are disabled. So now the dialog basically enters closing animation
state (cDialog.closing) when a button is pressed, except the update
callback then instantly completes it and calls the handler.

Looking into this bug was one of those cases where I just couldn't
reproduce the issue nor find any possibility for error in any of the
relevant code, and started suspecting it was some kind of LÖVE issue
tied to specific hardware or specific graphics drivers, or LÖVE falling
back to an alternative dependency (that doesn't properly handle drawing
non-power-of-two canvases to other canvases or whatever) because the
preferred library was missing or something like that. Anyway, glad it
was something explainable and fixable after all. Thanks to Misa for
reporting this.

[pre49]