Add decoding of Vorbis metadata for loop points (no actual looping yet)
Earlier I said supporting loop points in Ved might not be feasible, or
could be a huge mess, but... I looked into the Ogg and Vorbis formats,
and then I wanted to implement it, so here we are. The format is less
complicated than I was expecting, so the code isn't even really a mess.

The function supports approximately what current SDL_mixer supports.
The SDL_mixer version that's currently shipped with VVVVVV appears to
be a bit more outdated and supports LOOPSTART/LOOPLENGTH/LOOPEND, with
a value in samples. Current SDL_mixer has support for alternate names
like LOOP_START and LOOP-START, as well as (partial) HH:MM:SS.mmm
times... So Ved accepts these too.

This commit only adds support for reading the metadata, not yet for
actually looping audio. Speaking about actually looping audio, there's
two ways I can think of to do that in LÖVE: (1) continuously paying
attention to the current time of the audio Source in love.update() and
seeking to the start position on time; or (2) creating a SoundData with
just the looping part and, when the intro is done, stopping the full
song and continuing (hopefully smoothly) with the looping part.

My guess is that method 1 will frequently overshoot the end point and
may even be complicated depending on what LÖVE does with a song ends
(won't it reset the time to 0?), and that method 2 will use more memory
but may work just well enough. A method 3, in case stopping and
starting audio is too noticeable but seeking in the same audio isn't,
could be to make a SoundData that has the intro, the looping part, and
then again the first second or so of the looping part.

For 1.9.0 I'll probably go no further than add a visual indication to
the track to indicate the loop points (let's get my priorities
straight, I kinda got carried away with this one) and then in a later
version I'll probably implement method 2, and also do some more things
with this new metadata (like displaying the sample rate).

[pre50]