Make input.getpos() give visual position by default, add `virtual` arg
The X position returned by input.getpos(id) could, by design, give a
value beyond the end of the line, because if you go from a position on
one line to another and there's a small/empty line in the way, you
don't want your cursor to immediately stay attached to however long the
shortest line was. I'm calling that a "virtual" position because I
looked at Vim and it has a 'virtualedit' mode (where you can set your
cursor anywhere you want, even beyond the end of lines), and yeah, it's
a position that doesn't really exist.

This commit ensures that input.getpos(id) now gives a "normal" position
by default, so capped at the length of the line, and you only get that
"out of bounds" coordinate if you ask for it by setting the second
argument (`virtual`) to true. I went through all usages of the function
and checked whether each one would need it. This fixes several bugs,
all triggered if you go from a longer line to a shorter one:

- Pressing Ctrl+U/Ctrl+K/Ctrl+Backspace/Ctrl+Del to erase to the start/
  end of the line or to delete a word to the left or right would erase
  too much, or would erase stuff despite being at the end of the line
  and having nothing to erase
- "Java IDE behavior" in the script editor continued
- The "Column" counter in the script editor even indicated this wrong
  number, I posted a video on Discord and missed that myself
- Probably more

And as a bonus, some code in the input system now is a bit cleaner
since it doesn't need to do this end-of-line capping itself anymore.

[pre51]