Remove SourcePackages, Windows and Mac folders
It was never a good idea to put all the release packages on the repo,
and for a long time now I've been wanting to change it, but due to how
Git works, removing them wouldn't be that useful OR would need the
ENTIRE history to be rewritten, so I just kept doing it this way. Plus
it wasn't THAT bad... But let's now stop making the problem worse. All
things considered I still don't think it's worth it to "nuke" the
entire history to really clean up these files (unless other people
convince me, then I guess I might go ahead and do it), but it's still
useful to remove the folders via commit, so there's no confusion over
Ved being up-to-date because there's no newer ones in the folder.

The files for every release are now over here:
    https://tolp.nl/ved/files/download/

For now it's just a standard Apache file index (I mean, you can sort
the columns! I should mess around with the last modified dates though),
but I might make something custom later, and I'm also thinking of
adding a way to download every single version at once, now that it's no
longer possible to do that by just cloning/downloading the repo.