Make script editor warning lights red instead of yellow
I wanna take the opportunity to make certain things that people often
seem to be overlooking (especially the Ctrl/Shift hint text when
placing a script box or terminal) stand out more.

One such thing is the warning lights in the script editor; I've always
been wondering whether they catch the attention they need, but the
general idea was: yellow was meant for things that might not be a
problem, and red was reserved for things that were probably/definitely
a problem. But the current yellow may blend in a bit too much (hard to
test how well it catches the attention of people unfamiliar with the
interface when you designed it yourself and especially when you're
fully paying attention to the lights to see how they look... But
recently there was a conversation where people did seem to either miss
it or think nothing of it.)

I also can't imagine the current yellow warnings to show up really
often as false positives, the user can still ignore them, and there's
currently no actual support to distinguish between lights that should
be yellow and lights that should be red. So I'll just make them all
red. And to make them stand out even more, they now also flash a
*little* bit, enough to make you not think "that's probably nothing"
if you're unfamiliar with the lights.

[pre01]