Add icon in script editor for internal scripting mode being on/off
When I add internal scripts to a level, I sometimes forget to enable
internal scripting mode, and the result is that the script just doesn't
work (or does only part of what it's supposed to). The main problem
here is that the int.sc mode button is designed to catch your attention
when it's on (by lighting up yellow), but when it's off, it's just
quietly sitting there - so it's easy to be unaware of it being off.

So now I added an icon in the top left - which is a green script icon
if internal scripting is off, and a blue script icon with a cog if it's
on. So now both modes have a recognizable color and icon, where green
represents something easy, and the blue represents something more
difficult/intermediate.

[pre03]