Update Translation/tools/create_language_po.lua for split help pages
That tool is helpful in case there's a new translation made in lua, and
the whole thing needs to be converted to po so it can be imported into
Pootle. The converter hadn't been used for a while (usually translators
haven't started yet when they join so we can start in Pootle with a
clean slate) but there's now a new translation (see next commit) which
was made in lua. Making the first version in lua has the advantage that
you can immediately test your translations, so it's something I want to
support properly and I'll probably be asking new translators which way
they want to start (doing it in lua also means you have no progress
bars for example, especially since there's no easy way to see whether a
string is untranslated or has to be identical)

The tool still worked, except for plural forms and split help pages.
Plural forms are easy to copy-paste manually, but I don't want to
copy-paste 491 help paragraphs by hand (which makes me wonder if I
actually did do that for multiple languages when I first split the help
articles or if I just forgot where I left the tool for that). So now
each help article is also processed if the number of paragraphs is the
same. The tool also got some other updates.