Add support for loop points in music editor
This works with QueueableSources. I had to first experiment how exactly
the API for QueueableSource behaved (the documentation is a bit
incomplete for them), and this could still use some optimization,
particularly by only loading SoundData when actually starting to play
a song, but it works really well now!

The Source in each song's table (in music) is now replaced by a wrapper
class, cVedSource or cVedQueueableSource, which both behave like a
regular Source. I'm also considering extending cVedSource to at least
have rudimentary loop points support in LÖVE 0.9 and 0.10, and maybe
even to fix time indicators being broken and missing in 0.9, if
possible (and easy).

[pre02]