Make window creation/resizing smoother
There were more mode changes and "recreations" than necessary, both
when starting Ved and when changing the display settings. In the most
extreme case, with both scaling and "smaller window width" mode, the
window would be spawned four times when starting Ved:

- Initial 896x480 window creation due to love.conf
- Resizing to 800x480 due to smaller window mode
- Resizing to 1600x960 due to smaller window mode + 2x scale
- Repositioning to center the window

As for the last point, I've removed love.window.setPosition(...)
altogether, since it turns out to be unnecessary: it was only run after
love.window.setMode(...) to re-center the window after resizing.
However, it's already possible to specify X and Y coordinates in
love.window.setMode(...) directly. As for the other points, there's now
only a single love.window.setMode(...) in the code in total. Starting
Ved or changing display settings should feel a lot smoother now, since
the only "wasteful" window change is the initial 896x480 window, as set
up in love.conf. I'm also considering turning that off (which turns out
to be possible as well!) so we can create a correctly sized window the
first time around.

[pre03]