Always set window icon and title
The window icon and title used to be set in love.conf, with the last
changes the window would not have a title until love.update first ran
and the window would not have an icon until you go to the main editor
(where the icon changes into the current tool's icon). Now the icon and
title is consistently initialized, and always has the version number.

I also dropped the "v" from the version number in the title - i.e.
"Ved 1.9.2" instead of "Ved v1.9.2".

[pre05]